Move density structures to subpackage
