+++ src/lintsampler/density_structures/base.py
