+++ src/lintsampler/density_structures/grid.py
@@ -1,8 +1,8 @@
 import numpy as np
 from functools import reduce
 from warnings import warn
-from .utils import _is_1D_iterable, _choice
-from .density_structure import DensityStructure
+from ..utils import _is_1D_iterable, _choice
+from .base import DensityStructure
 
 
 class DensityGrid(DensityStructure):
