+++ src/lintsampler/lintsampler.py
@@ -1,8 +1,8 @@
 import numpy as np
 from warnings import warn
 from scipy.stats.qmc import QMCEngine, Sobol
-from .density_grid import DensityGrid
-from .density_structure import DensityStructure
+from .density_structures.grid import DensityGrid
+from .density_structures.base import DensityStructure
 from .utils import _is_1D_iterable, _choice, _check_hyperbox_overlap, _all_are_instances
 from .sampling import _grid_sample
